Updated simulation algorithm.

+++ dev/trunk/simmodel/src/main/java/edu/trafficsim/model/simulator/DefaultSimulator.java
@@ -9,12 +9,12 @@ public class DefaultSimulator extends AbstractSimulator<DefaultSimulator>
 
 	private static final long DEFAULT_SEED = 0l;
 
-	public DefaultSimulator(long id, String name, double duration,
+	public DefaultSimulator(long id, String name, int duration,
 			double stepSize) {
 		this(id, name, duration, stepSize, DEFAULT_SEED);
 	}
 
-	public DefaultSimulator(long id, String name, double duration,
+	public DefaultSimulator(long id, String name, int duration,
 			double stepSize, long seed) {
 		super(id, name, duration, stepSize, seed);
 	}
